model/requirements.txt
python-dotenv
spacy
ultralytics
opencv-python
cvzone
numpy
scipy
scikit-learn
pandas
torch
firebase-admin
transformers
schedule